Refactor: 엔티티의 필요없는 필드 삭제
Budget, Expense 엔티티의 userId 필드 삭제
