+++ src/main/java/com/saving/category/domain/repository/CategoryRepository.java
@@ -7,6 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     List<Category> findAllByUserId(Long userId);
-
-    boolean existsByIdAndUserId(Long id, Long userId);
 }
