+++ src/main/java/com/saving/expense/domain/entity/Expense.java
@@ -1,7 +1,8 @@
-package com.saving.expense.domain;
+package com.saving.expense.domain.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Builder;
@@ -15,22 +16,20 @@ import lombok.NoArgsConstructor;
 public class Expense {
 
     @Id
-    @GeneratedValue
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long userId;
-    private String categoryName;
-    private char expenseMethod;
+    private Long categoryId;
+    private ExpenseMethod expenseMethod;
     private int amount;
     private String content;
     private int isTotalExpenseApply;
     private String expenseAt;
 
     @Builder
-    public Expense(Long userId, String categoryName, char expenseMethod, int amount,
+    public Expense(Long categoryId, String expenseMethod, int amount,
             String content, int isTotalExpenseApply, String expenseAt) {
-        this.userId = userId;
-        this.categoryName = categoryName;
-        this.expenseMethod = expenseMethod;
+        this.categoryId = categoryId;
+        this.expenseMethod = ExpenseMethod.of(expenseMethod);
         this.amount = amount;
         this.content = content;
         this.isTotalExpenseApply = isTotalExpenseApply;
